Renames 'healthz' to 'health'
