+++ api/health/health.go
@@ -1,4 +1,4 @@
-package healthz
+package health
 
 import (
 	"encoding/json"
@@ -7,18 +7,18 @@ import (
 	"github.com/cloudflare/cfssl/api"
 )
 
-type HealthzResponse struct {
+type HealthResponse struct {
 	Healthy bool `json:"healthy"`
 }
 
-func healthzHandler(w http.ResponseWriter, r *http.Request) error {
-	response := api.NewSuccessResponse(&HealthzResponse{Healthy: true})
+func healthHandler(w http.ResponseWriter, r *http.Request) error {
+	response := api.NewSuccessResponse(&HealthResponse{Healthy: true})
 	return json.NewEncoder(w).Encode(response)
 }
 
 func NewHealthCheck() http.Handler {
 	return api.HTTPHandler{
-		Handler: api.HandlerFunc(healthzHandler),
+		Handler: api.HandlerFunc(healthHandler),
 		Methods: []string{"GET"},
 	}
 }
